WIP: Adjusted Text Tests
